MuPDF: Experimental interface
Playing around with using MuPDF instead of Poppler to render pages:

  http://mupdf.com

MuPDF is written in C, so I have added a C++ header, `MuPDF.hpp`, that wraps
the C functions into a C++ class that can hopefully be used in a similar
fashion to Poppler's `Poppler::Document`. Currently the `MuPDF::Document` just
loads a filename and initializes top-level MuPDF objects.

`PDFDocumentScene` has been hacked to hold a reference to `MuPDF::Document` in
parallel with `Poppler::Document`. The goal is to replace page rendering with
something based on MuPDF in order to assess efficiency.

If things look good, may explore further.